+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [2.0.0-beta2] - 2024-11-28
+
+### Added
+
+- New `BerliozCommandJobHandler` handler to execute Berlioz command
+- New `BerliozSystemJobHandler` handler to execute shell command
+
 ## [2.0.0-beta1] - 2024-11-25
 
 Start development.